// smsf-masterdata-ui/smsf-xchange-ui/src/app/services/rest/model/xChangeDTO.ts
/**
 * Automation API provider
 * Automation provides template of this description, and anyone may modify it in your own project.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BusinessCaseDTO } from './businessCaseDTO';
import { BusinessDataDTO } from './businessDataDTO';


export interface XChangeDTO { 
    businessCase?: BusinessCaseDTO;
    businessDataList?: Array<BusinessDataDTO>;
}